requirements.txt
fastapi
uvicorn
python-dotenv
stripe
jinja2
python-multipart
sqlalchemy
psycopg[binary]
pydantic-settings
celery
redis
openai
email-validator
numpy
opencv-python-headless
mediapipe
